better handling of get target state errors
This patch improves handling of get target state errors. When we
get an error, we still want to run the main reconciler, because
the controller relies on periodic updates from operator. So, instead
of returning empty from the poller, we return an updated status
with an empty target state - which forces the framework to re-run
reconciliation.