fix: initialize TTDMockAdmin with the changelogs to create up front (#124)
The main fix here is in how we "create" topics within the MockAdmin used by the RTTD. Because no real topics are created in the TTD context, everything has to mock its own internal view of the "cluster" which in the admin case, means making sure that it adds any topics that we will later make requests for. For us, this is just changelogs (and should generally only ever be changelogs, as the input topics are completely abstracted away by the TTD entirely including everything like the KafkaClientSupplier)

Before this, we would lazily create topics when requested for in the #describeTopics API. However for FACT tables we skip this particular call, and so incur an error when we next hit the mock admin in the #describeConfigs call.

The fix is to remove any topic-creation inside Admin overrides (since this is extremely fragile and relies on us either overriding every method that is ever added to it or else being super diligent about keeping the TTDMockAdmin up to date with any admin calls ever added to our codebase). Instead, we pass in the appId and topology to the TTDMockAdmin's constructor so that it can derive the changelog topic names from the list of state store names, and add everything up front. Note that we can't get a handle on the changelog topic names directly until after the topology is initialized, and unfortunately this does not occur until we're already inside the TopologyTestDriver constructor, where the exception in this bug originates.

However, as a bit of an added safety measure, I also added a check after we exit the TTD constructor and can make sure that all the changelogs were created as expected. This doesn't really do much, but it'll be nice to have as a warning in case something does change that invalidates any assumption in this fix