Finalize ResponsiveKafkaStreams API and clean up StreamsConfig (#136)
Get this ready to publish the API Reference docs. Fills in a few niche missing constructors (KafkaStreams allows you to pass in a Time object), adds javadocs for everything, and cleans up the way we handle StreamsConfigs and validate the properties.

Introduces a new ResponsiveStreamsConfig class so we can do "quiet logging". Will follow up with the same for the other kafka clients that we're excessively logging