Remove timestamped store suppliers and ResponsiveKafkaStreams#create (#135)
The main changes are just removing the timestamped store implementations, and the static constructor ResponsiveKafkaStreams#create which is no longer needed. The exact steps to migrate are listed below, but to clarify, the reason for the change is to simplify the API and make it as easy as possible to understand and implement. We can remove the timestamped stuff because, for non-persistent stores, Streams will basically do this simple wrapping and unwrapping automatically and we don't need the marker interface. Note that this does require that we never return true from the stores' #persistent API, as Streams infers this to mean "RocksDB specifically" and does some weird stuff for the timestamped stores.

This is a breaking change for our active users. Migration steps after this PR:

1. Do a find and replace of ResponsiveKafkaStreams.create to new responsiveKafkaStreams
2. Replace the timestamped StoreSupplier with just the plain one. This is also a simple find-and-replace, though it depends on which version of the store supplier API they were using before:
i. ResponsiveKeyValueParams.timestamped becomes ResponsiveKeyValueParams.keyValue
ii. ResponsiveKeyValueParams.timestampedFact becomes ResponsiveKeyValueParams.fact
iii. ResponsiveStores.timestampedKeyValueStore becomes ResponsiveStores.keyValueStore

All of the above can be achieved by just running the following commands on their application code file

> sed -ie "s|ResponsiveKafkaStreams.create|new ResponsiveKafkaStreams|g" ${FILE}
> sed -ie "s|ResponsiveKeyValueParams.timestamped|ResponsiveKeyValueParams.keyValue|g" ${FILE}
> sed -ie "s|ResponsiveKeyValueParams.timestampedFact|ResponsiveKeyValueParams.fact|g" ${FILE}
> sed -ie "s|ResponsiveStores.timestampedKeyValueStore|ResponsiveStores.keyValueStore|g" ${FILE}