Log version and git commit id during startup (#142)
Added a task to the gradle build that stashes some various useful git properties in a file along with the main build. We can then read this out to log valuable debugging info such as the version and specific commit id of the dependency.

Tested this out both in a local kafka-client integration test, and by pulling in a local build with the fix while running a test in the soak repo. Works with both, and logs for the TTD as well as real applications.

Useful for sanity checks, both in the soak test and future customer logs where we want to verify exactly what version they're running. I only put it in the kafka-client module since that's where it seems most relevant, but we could extend it to others as well if that feels right.