Consolidate metrics and group by scope (#156)
The primary goal here is to consolidate all of the metrics added by us and then group them according to their scope, so that we can attach a set of common base tags to all metrics with additional tags added based on the specific context

To help track tags for a given scope, I added a MetricGroup interface with three starter implementations: ApplicationMetrics, TopicMetrics, and StoreMetrics. The first two are already in use in this PR as I migrated the existing metrics which should help demonstrate the approach.