Maintain the shared session clients as a singular container when passed via internal configs (#166)
Relatively small in scope refactoring to change the way the SharedClients class is handled. Right now we assemble and disassemble the entire SharedClients every time we need to access something from it, regardless of whether we in fact needed the entire object or just a single one of its clients. This also ensures that something added to or modified in the SharedClients container is percolated throughout the application (for example if we need to add another client after going through the KafkaStreams constructor, which is currently impossible to do).

Also renames SharedClients to SessionClients