More application health metrics: restoration and thread errors (#167)
It'd be nice to be able to confirm that our tasks are in fact spending little (or no) time in restoration. This is less about verifying the fundamental hypothesis of remote stores and more about general monitoring of application stability. Unlike with plain Kafka Streams, a high value for any of these metrics in one of our applications would indicate an unhealthy state and should be considered a signal to investigate

To that end, this PR adds three new Responsive metrics using the custom restore listener and one existing Streams metric for the number of "failed threads" (ie thread death & replacement events)

1. num-restoring-changelogs -- simple count of stores currently in restoration
2. num-interrupted-changelogs -- counter for stores being interrupted before finishing restoration
3. restore-latency -- how long a store spent being restored from its changelog
4. failed-stream-threads -- how many times has a thread been killed and replaced on this client