Remove batchSize parameter from writers (#179)
Relatively small refactor to clear up how batches are handled/sized between the CommitBuffer and RemoteWriters. It seems that the batchSize we pass in is always either the MAX_BATCH_SIZE constant or, in the case of restoration only, the size of the final batch that contains whatever records are remaining up to the MAX_BATCH_SIZE

In other words, the batch size is always capped at MAX_BATCH_SIZE and furthermore, the only time the batchSize parameter is different than the hard-coded constant is when there are fewer updates in the batch and thus the batch will be naturally capped at this smaller size since we will run out of inserts (or deletes) before hitting the batchSize condition. Also, only the lwt writer seems to be using this at the moment, so it feels silly for us to pass this in to every other writer as we do right now.

In addition to just being easier to reason about and removing unnecessarily complexity from the code, this is a necessary cleanup/precursor for the window store implementation