Use constant collection names and validate TableName length (#213)
Mongo limits database names to 65 characters (at the most -- may be even smaller in some cases) and also enforces the combined <database_name>.<collection_name> string to not exceed 128 characters.

We should validate the table name length up front for fast-failure (I actually ran into this in a test with a long name). We should also remove the table name from the collection naming scheme for the MongoKVTable. We don't need to include the table name again since the collections are by definition scoped to that database, which is named after the state store. This way we are able to use up to the full allotted character count for table names (since otherwise the metadata collections could exceed 128 characters with a 64-character table name).