first pass at supporting migrating by setting a config
This is a first pass at supporting migrating by setting a config.
When the config is on, ResponsiveKafkaStreams runs migration instead
of the application.

The migration requires discovering the set of stores. This is maintained
in a singleton thats updated by the store supplier constructor.

Currently we only support a single non-source store