couple tweaks to support safe concurrent access to stores
This patch adds support for concurrent access to stores. After this change, we should
be able to concurrently issue read calls to stores while other reads are happening,
or while the main stream thread is doing put/delete or flushing.

First, we add a mechanism for injecting values into the store that can only be computed
after the store has been initialized. We need to do this because the stores depend on
some info that is made available only from the call to , which is kind of a
quirk in streams (ideally, these things are just parameters to put/get/etc). In particular
in today's implementation we depend on the processor context supplied in init to get the
record's timestamp which we need for expiry calculation. This patch supports reading
this from a reference to  in the store registration. When an async
getter is reading the store, it can inject the appropriate context.

Second, we use a concurrent map for the commit buffer's store of uncommitted records