Use TimestampedKeyValueStore in AsyncProcessorIntegrationTest (#280)
Forgot to merge this with the async POC branch before merging that PR, so here's a quick followup that makes sure we actually use the TimestampedKeyValueStore in the async integration test.

I needed to make a few small changes to the way we build the normal (non-async) state stores for this to work, most notably adding a ResponsiveTimestampedKeyValueStore that implements the TimestampedBytesStore marker interface in Streams. We actually had this a while back but I took it out to simplify things because we didn't strictly need it for plain Streams. Well, now we need it for the async processing stuff.

(If you're curious why we need it now, it's because we override the actual builder that adds the various store layers so that we can stick in our special async flushing layer to the async stores. And the Streams builder we override does some fancy magic to wrap stores that don't implement the TimestampedBytesStore in a timestamping-layer. So we could either re-implement this fancy/hacky magic in our builder, which is a non-trivial amount of extra code, or we could just have an actual timestamped version of our own KV store, which is actually very little extra code since it just has to add the marker interface but can otherwise simply extend the non-timestamped store. Also it's probably the right thing to do anyway)