model mongodb keyvalue keys as strings to support range scans (#314)
* add an integration test for foreign key joins

* model mongodb keyvalue keys as strings to support range scans

This patch changes the mongodb key representation to a string from
a byte array. The string is a base64 encoded representation of the
key bytes. The specific encoding uses a character table that
guarantess that the byte-order of the keys is preserved in the
mongo string sort order.

This change is required to support range scans in mongodb. MongoDB
orders binary fields first by length, and then by content for fields
of the same length. This is not the order required/assumed by
Kafka Streams.

There is likely a performance and space penalty for this change,
as the base64 encoded keys are 8/6 the size of the binary keys.
If its a problem, then for use cases that do not require range scans
we can add a config that toggles the key representation. However
I don't think its worth the added complexity just yet. We can revisit
if it actually manifests as a problem.

Extending this change to other store types is deferred to a follow-on
patch.

* add test

* use byte/byteArray methods in test

* fix rks integration test

* checkstyle

---------

Co-authored-by: Rohan Desai <rohan@responsive.dev>