Row-level TTL PR 7: TopologyTestDriver#advanceWallClockTime (#379)
Minor followup to PR #6 where we changed the TTD so that only advancing stream-time by piping additional input records would apply ttl. Although this more closely matches how the TTD has to be used in vanilla Streams, it's a pain to be unable to advance the time on its own without being coupled to the input records. This PR restores the previous semantics of the ResponsiveTopologyTestDriver#advanceWallClockTime API so that this can be used to validate ttl in TTD tests