add support types for license verification (#385)
* add support types for license verification

This patch adds some infra to support license verification:

The main change is the addition of LicenseAuthenticator and
LicenseChecker. LicenseAuthenticator is responsible for taking
a license document, verifying that it is signed correctly, and
then returning the contained signed LicenseInfo. LicenseChecker
takes a LicenseInfo and validates that the license it describes
can be used.

These new classes operate on the pojo types defined under the model
package that define the structure of the license documents, and the
key registry (SigningKeys) that's to be packaged into the jar.

Finally, this patch also includes a utility class for parsing pem
files, which we'll use to store public keys. I opted to write our
own parser (rather than using something like bouncycastle) because
the format it's simple and I'd rather not include a dependency on
a foundational artifact like bouncycastle that is likely to
conflict with a user's dependencies.

* fix string replace