Change license loading precedence (#450)
Currently we will not load a license if the API key is provided through `responsive.platform.api.key`. This patch changes the precedence to use this key as a fallback only if license loading fails. 

Additionally, we change RS3 initialization logic to not pass through an API key when the license is not `UsageBasedV1`.